Separate arg parsing and use rollup for conversion to ESM
In order to facilitate incorporation in the LVGL Pro codebase
