+++ .gitignore
@@ -7,3 +7,5 @@ dist
 *.log
 *.swp
 .vscode
+esm_output
+**/.DS_Store
