+++ web/convert_browser.js
@@ -2,6 +2,8 @@ const collect_font_data = require('../lib/collect_font_data');
 const binWriter = require('../lib/writers/bin');
 const lvglWriter = require('../lib/writers/lvgl');
 
+const { parse_args } = require('../lib/parse_args');
+
 const writers = {
   bin: binWriter,
   lvgl: lvglWriter
@@ -14,5 +16,5 @@ async function convertBrowser(args) {
   return files;
 }
 
-module.exports = convertBrowser;
+module.exports = { convertBrowser, parse_args };
 module.exports.formats = Object.keys(writers);
